Fix strncmp checking for i2c
Lukanite change

Changes strncmp to memcmp because the first byte in pn532response_firmwarevers is 0x00, which causes strncmp to prematurely return as a match as long as both it and pn532_packetbuffer start with 0x00.

In the original code when using I2C, a 1-byte offset causes a mismatch between pn532response_firmwarevers and pn532_packetbuffer, but strncmp fails to catch this and continues to proceed.

readack() has also been updated for a similar reason.

Scope: Firmware version checking, Ack reading

Known limitations: None

Tests: Run firmware version checking (like in readMifare) in both I2C and SPI modes